Add basic navbar loading
This adds a basic navbar loader, which will show either
login or logout depending on if the user is logged in or
not.

I choose to use session storage to actually store the info
rather than cookies because it's a much simpler interface,
but I use cookies for the protocol since it plays slightly nicer
with FastAPI auto-generated docs compared with custom headers.