Add sign-up logic
Currently the alerts never go which is a bit more accessible but will
likely be annoying on pages which we expect multiple alerts on.
We might be able to get away with non-alert notifications on those
pages, however.

If we do add dismissing alerts, they will need a transition which is
somewhat complicated