Add favicon, get rid of 302 redirects
Now everything can get loaded directly from the disk cache
or memory cache. The few things that could use an update
will use a stale-while-revalidate cache control header.

The idea is that after viewing a page you could turn off your internet
and reload and it would still load for a while. This won't be possible
on the login / signup pages due to recaptcha.

This will prevent page-view style metrics from being collected but
that's a perfectly ok trade for now.