Add favicon, improve performance, add logs page (#4)
* Add favicon, get rid of 302 redirects

Now everything can get loaded directly from the disk cache
or memory cache. The few things that could use an update
will use a stale-while-revalidate cache control header.

The idea is that after viewing a page you could turn off your internet
and reload and it would still load for a while. This won't be possible
on the login / signup pages due to recaptcha.

This will prevent page-view style metrics from being collected but
that's a perfectly ok trade for now.

* Add navbar permissions checks, logs in navbar

* Add VS code settings to hide ico/png files

They just clutter stuff up

* Add logging page

The logging page is everything I was hoping it would be. Tail-mode which
isn't horribly inefficient, text filters, start-at dates, and is
extremely mobile-friendly