Add initial response rendering
This doesn't do the AJAX calls yet; I'm imagining that the response
histories are paginated. Furthermore, I think it's going to be very
helpful if we add "response_substitutions"or something
along those lines for all the values which can be substituted on
a given response so it can be validated.

This is going to also need some pagination on the response histories
and it won't actually load any of the response histories for a
response unless you go to preview it and ask, as that would be a huge
waste