Add responses page (#11)
* Add initial response rendering

This doesn't do the AJAX calls yet; I'm imagining that the response
histories are paginated. Furthermore, I think it's going to be very
helpful if we add "response_substitutions"or something
along those lines for all the values which can be substituted on
a given response so it can be validated.

This is going to also need some pagination on the response histories
and it won't actually load any of the response histories for a
response unless you go to preview it and ask, as that would be a huge
waste

* Visually add response descriptions

I think response descriptions will be worth it to add here; they can
be used to document the allowed substitutions without being as
complicated as a a data-backed substitution model

* Add initial ajax flow methodology

* Refactor slide animations

This is going to need further refactoring, but HeightEased will
support adding a height easing onto any component. It's...
not the most elegant solution

* Finish up responses area

This was not the cleanest section, but at least it's admin-only and
unlikely to ever change. I will put these lessons into the next sections
rather than refactoring this, at least for now

* Make edit reasons visible