Progress on loan summary
It took a ton of finicking around to get 3 distinguishable colors
that made sense for "repaid, unpaid, inprogress". I started out with
the intention of blue/green/red but the devil is in the details. The
final realization was that I really needed blue-green for inprogress,
not really blue.

My tests with NoCoffee Vision Simulator suggest this should be
distinguishable by everyone